requirements.txt
streamlit
faiss-cpu
sentence-transformers
pymupdf
openai
numpy
pandas